+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 This project attempts to conform to [SemVer](https://semver.org/) guidelines whenever possible.
 
 ## [v2.0.0]
+* [Looking up Minecraft usernames for the Shop commands](https://github.com/Ubunfu/mc-steve/pull/46)
 
 ## [v1.11.0]
 * [Modularly including purchase and/or sell price in get item response](https://github.com/Ubunfu/mc-steve/pull/44)